+++ benchmark/reports/bytecode_size_comparison.md
@@ -1,6 +1,6 @@
 # Bytecode Size Comparison
 
-Generated on: 2026-01-26
+Generated on: 2026-01-27
 
 ## Fibonacci Implementations
 
@@ -22,14 +22,12 @@ Generated on: 2026-01-26
 | SimpleToken_pvm            | PVM     | 5,275        | +797.1%     |
 | simple_token_ink           | PVM     | 7,251        | +1133.2%    |
 
-
 ## TetherToken Implementations
 
 | Contract        | VM Type | Size (bytes) | vs Smallest |
 | --------------- | ------- | ------------ | ----------- |
 | TetherToken_evm | EVM     | 5,860        | -           |
 | TetherToken_pvm | PVM     | 53,640       | +815.4%     |
-
 
 ## WETH9 Implementations
 
@@ -38,14 +36,12 @@ Generated on: 2026-01-26
 | WETH9_evm | EVM     | 2,204        | -           |
 | WETH9_pvm | PVM     | 17,712       | +703.6%     |
 
-
 ## FiatTokenV2_2 Implementations
 
 | Contract          | VM Type | Size (bytes) | vs Smallest |
 | ----------------- | ------- | ------------ | ----------- |
 | FiatTokenV2_2_evm | EVM     | 15,400       | -           |
 | FiatTokenV2_2_pvm | PVM     | 113,736      | +638.5%     |
-
 
 ## FiatTokenProxy Implementations
 
@@ -54,7 +50,6 @@ Generated on: 2026-01-26
 | FiatTokenProxy_evm | EVM     | 1,836        | -           |
 | FiatTokenProxy_pvm | PVM     | 9,882        | +438.2%     |
 
-
 ## XENCrypto Implementations
 
 | Contract      | VM Type | Size (bytes) | vs Smallest |
@@ -62,12 +57,9 @@ Generated on: 2026-01-26
 | XENCrypto_evm | EVM     | 10,983       | -           |
 | XENCrypto_pvm | PVM     | 101,780      | +826.7%     |
 
-
 ## CoinTool_App Implementations
 
 | Contract         | VM Type | Size (bytes) | vs Smallest |
 | ---------------- | ------- | ------------ | ----------- |
 | CoinTool_App_evm | EVM     | 4,661        | -           |
 | CoinTool_App_pvm | PVM     | 32,680       | +601.1%     |
-
-
